refactor: 이름 검증 로직 Name 클래스로 이동
